--- src/App.css
@@ -1,131 +1,0 @@
-.App{
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-   width:80%;
-    margin: 0 auto;
-    padding-top:200px;
-}
-*{
-    margin: 0;
-    padding: 0;
-}
-.prepare .content{
-    box-sizing: border-box;
-    padding: 10px 20px 10px;
-    background-color: pink;
-    height: 400px;
-    width: 200px;
-    text-align: center;
-}
-.prepare .title{
-    background-color: pink;
-    width: 200px;
-    text-align: center;
-    display: inline-block;
-    margin-bottom: 20px;
-    line-height: 40px
-}
-.learning .content{
-    box-sizing: border-box;
-    padding: 10px 20px 10px;
-    background-color: #7aff98;
-    height: 400px;
-    width: 200px;
-    text-align: center;
-}
-.learning span{
-    background-color: #7aff98;
-    width: 200px;
-    text-align: center;
-    display: inline-block;
-    margin-bottom: 20px;
-    line-height: 40px;
-}
-.complete .content{
-    box-sizing: border-box;
-    padding: 10px 20px 10px;
-   background-color:#eee;
-    height: 400px;
-    width: 200px;
-}
-.complete span {
-    background-color: #eee;
-    width: 200px;
-    text-align: center;
-    display: inline-block;
-    margin-bottom: 20px;
-    line-height: 40px;
-}
-.addContent{
-    display: inline-block;
-    border-radius:50%;
-    font-size: 40px;
-    width: 50px!important;
-    height: 40px;
-    margin-bottom: 0!important;
-    margin-top: 20px;
- border: 2px solid #eeeeee;
-    color: #eeeeee;
-}
-.contentItem{
-    padding: 5px 10px;
-    border: 2px solid black;
-     display: flex;
-    justify-content: center;
-    align-items: center;
-    box-sizing:border-box;
-    margin-top: 10px;
-    /*text-overflow: ellipsis;*/
-    /*overflow: hidden;*/
-    /*white-space: nowrap;*/
-    height:40px;
-}
-.focus{
-    height:34px;
-    margin-top: 10px;
-    width: 157px;
-    border: 2px solid black;
-}
-.hide{
-    display: none !important;
-}
-.inputBlur{
-    height: 34px;
-    line-height:34px;
-    margin-top: 10px;
-    width: 157px!important;
-    display: block;
-    border: 2px solid black;
-}
-.isEnter{
-    border: 1px solid red;
-}
-.delete::after{
-    position: relative;
-    left: 74px;
-    top:-58px;
-    font-size:20px;
-    border-radius: 50%;
-    background-color: red;
-    width: 30px;
-    height: 30px;
-    display: inline-block;
-    text-align: center;
-    color: #eeeeee;
-    content:"x";
-}
-.delete2::after{
-    position: relative;
-    left: 150px;
-    top:-58px;
-    font-size:20px;
-    border-radius: 50%;
-    background-color: red;
-    width: 30px;
-    height: 30px;
-    display: inline-block;
-    text-align: center;
-    color: #eeeeee;
-    content:"x";
-}
