+++ src/mock/login.json
@@ -1,0 +1,4 @@
+{
+  "username":"admin",
+    "password": "123456"
+}
