+++ src/views/login/index.css
@@ -1,0 +1,27 @@
+button{
+    padding: 10px 78px;
+    border: none;
+    background-color: blue;
+    color: white;
+}
+.login{
+    transform: translate(-50%,-50%);
+    position: absolute;
+    left:50%;
+    top:50%;
+}
+.input{
+    border: 1px solid #7aff98;
+    margin-bottom: 30px;
+    padding: 5px 10px;
+    outline: none;
+}
+.message{
+    background-color: #7aff98;
+    width: 250px;
+    line-height: 50px;
+    color: green;
+    margin: 0 auto;
+    font-size: 14px;
+}
+
